Another huge update. Someday I will stop doing this.
The structure of the project have changed. Before, storage was unused, but now after parsing data, all data must be commited in to the storage.

Also, list model has moved to CPP code (it shows better performance and it allow to avoid mistakes with scripts)

New approach to request additional data - instead of one hude custom handler now it accumulate all simillar handlers in one and then request

This version may be called 0.3 Alpha

+++ src/vklongpollupdateparser.cpp
@@ -46,13 +46,13 @@ int VKLongPollUpdateParser::parseIds(const QString &data, int position, VKLongPo
         
         QString userId = "1";
         position = parseUserId(data, position, userId);
-        c.setUserId(userId);
-        m_uniqueUsers.append(userId);
+        c.setUserId(userId.toInt());
+        m_uniqueUsers.append(userId.toInt());
 
 
         QString msgId = "1";
         position = parseMsgId(data, position + 1, msgId);
-        c.setMessageId(msgId);
+        c.setMessageId(msgId.toInt());
 
         if (position == data.length()) {
             qDebug()<<"end1 add to"<<fwd.userId()<<c.userId();
